rename once-cell file to lazy-initialized
