+++ src/common/utils/lazy-initialized.ts
